Refactor Anchor component to utils folder.
